+++ src/items/EdgeItem.h
@@ -11,6 +11,11 @@ enum class EdgeRoutingMode {
     ObstacleAvoiding
 };
 
+enum class EdgeBundlePolicy {
+    Centered,
+    Directional
+};
+
 class EdgeItem : public QGraphicsPathItem {
 public:
     explicit EdgeItem(const QString& edgeId, PortItem* sourcePort, QGraphicsItem* parent = nullptr);
@@ -20,10 +25,14 @@ public:
     PortItem* sourcePort() const;
     PortItem* targetPort() const;
     EdgeRoutingMode routingMode() const;
+    EdgeBundlePolicy bundlePolicy() const;
+    qreal bundleSpacing() const;
 
     void setTargetPort(PortItem* port);
     void setPreviewEnd(const QPointF& scenePos);
     void setRoutingMode(EdgeRoutingMode mode);
+    void setBundlePolicy(EdgeBundlePolicy policy);
+    void setBundleSpacing(qreal spacing);
     void updatePath();
 
 private:
@@ -32,4 +41,6 @@ private:
     QPointer<PortItem> m_targetPort;
     QPointF m_previewEnd;
     EdgeRoutingMode m_routingMode = EdgeRoutingMode::Manhattan;
+    EdgeBundlePolicy m_bundlePolicy = EdgeBundlePolicy::Centered;
+    qreal m_bundleSpacing = 18.0;
 };